requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
beautifulsoup4==4.12.3
google-generativeai==0.8.0
